requirements.txt
ipykernel
python-dotenv
psycopg2-binary
requests
beautifulsoup4
lxml